hashrate and shares seem to never have decimals, adjusted code. Adjusted the names of the columns to match closer to the JSON API.

+++ src/net/nexustools/jbitminingstats/util/MiningWorkerStub.java
@@ -3,11 +3,11 @@ package net.nexustools.jbitminingstats.util;
 public class MiningWorkerStub {
 	public String name;
 	public boolean online;
-	public double hashRate;
-	public double share;
+	public int hashRate;
+	public int share;
 	public double score;
 	
-	public MiningWorkerStub(String name, boolean online, double hashRate, double share, double score) {
+	public MiningWorkerStub(String name, boolean online, int hashRate, int share, double score) {
 		this.name = name;
 		this.online = online;
 		this.hashRate = hashRate;
